refactor: extract file processing code from oci.py
This commit cleans up oci.py by removing all functions related to
processing files and directories for new OCI layers and placing them in
a new layers.py module. This makes sense because the code is largely
OCI-independent (working off of filesystem directories) and this gives
us a cleaner separation of concerns.

The corresponding unit tests are also removed from test_oci.py and
placed in the new test_layers.py module. There are no functional
changes, but the tests themselves are lightly updated to reflect the
fact that archive_layer() is a free function.

Fixes #296
