+++ tests/unit/test_oci.py
@@ -13,8 +13,6 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-# pylint: disable=too-many-lines
-
 import datetime
 import hashlib
 import json
@@ -22,11 +20,9 @@ import os
 import tarfile
 from collections import namedtuple
 from pathlib import Path
-from typing import List, Tuple
 from unittest.mock import ANY, call, mock_open, patch
 
 import pytest
-from craft_parts.overlays import overlays
 
 import tests
 from rockcraft import errors, oci
@@ -52,7 +48,7 @@ def mock_run(mocker):
 
 @pytest.fixture
 def mock_archive_layer(mocker):
-    yield mocker.patch("rockcraft.oci._archive_layer")
+    yield mocker.patch("rockcraft.layers.archive_layer")
 
 
 @pytest.fixture
@@ -93,27 +89,6 @@ def mock_write_bytes(mocker):
 @pytest.fixture
 def mock_add_layer(mocker):
     yield mocker.patch("rockcraft.oci.Image.add_layer")
-
-
-@pytest.fixture
-def temp_tar_contents(mocker) -> List[str]:
-    """Fixture that mocks `oci._add_layer_into_image()` to inspect the temporary
-    tarfile that is generated by `oci._archive_layer()`.
-
-    :return: A list with the names of the files/directories in the "intercepted"
-        tarfile.
-    """
-    contents = []
-
-    def list_tar(_image_path: Path, archived_content: Path, **_kwargs):
-        with tarfile.open(archived_content, "r") as tar_file:
-            contents.extend(tar_file.getnames())
-
-    mocker.patch.object(
-        oci, "_add_layer_into_image", side_effect=list_tar, autospec=True
-    )
-
-    return contents
 
 
 @tests.linux_only
@@ -329,296 +304,6 @@ class TestImage:
             call(expected_cmd + ["--history.created_by", " ".join(expected_cmd)])
         ]
 
-    def test_add_layer_directories(self, tmp_path, temp_tar_contents):
-        """Test that adding a directory as a layer explicitly preserves subdirs"""
-        dest_dir = tmp_path / "dest"
-        dest_dir.mkdir()
-
-        layer_dir = tmp_path / "layer_dir"
-        layer_dir.mkdir()
-
-        (layer_dir / "first").mkdir()
-        (layer_dir / "first/first.txt").touch()
-
-        (layer_dir / "second").mkdir()
-        (layer_dir / "second/second.txt").touch()
-
-        image = oci.Image("a:b", dest_dir)
-
-        assert len(temp_tar_contents) == 0
-        image.add_layer("tag", layer_dir)
-
-        expected_tar_contents = [
-            "first",
-            "first/first.txt",
-            "second",
-            "second/second.txt",
-        ]
-        assert temp_tar_contents == expected_tar_contents
-
-    def test_add_layer_symlinks(self, tmp_path, temp_tar_contents):
-        """
-        Test creating a new layer with symlinks (both file and dir).
-        """
-        dest_dir = tmp_path / "dest"
-        dest_dir.mkdir()
-
-        layer_dir = tmp_path / "layer_dir"
-        layer_dir.mkdir()
-
-        (layer_dir / "first_dir").mkdir()
-        (layer_dir / "first_file").touch()
-
-        (layer_dir / "second_dir").symlink_to("first_dir", target_is_directory=True)
-        (layer_dir / "second_file").symlink_to("first_file", target_is_directory=False)
-
-        image = oci.Image("a:b", dest_dir)
-
-        assert len(temp_tar_contents) == 0
-        image.add_layer("tag", layer_dir)
-
-        expected_tar_contents = [
-            "first_dir",
-            "first_file",
-            "second_dir",
-            "second_file",
-        ]
-        assert temp_tar_contents == expected_tar_contents
-
-    def test_add_layer_with_base_layer_dir(self, tmp_path, temp_tar_contents):
-        """Test creating a layer with a base layer dir for reference."""
-        dest_dir = tmp_path / "dest"
-        dest_dir.mkdir()
-
-        layer_dir = tmp_path / "layer_dir"
-        layer_dir.mkdir()
-
-        (layer_dir / "first").mkdir()
-        (layer_dir / "first/first.txt").touch()
-
-        (layer_dir / "second").mkdir()
-        (layer_dir / "second/second.txt").touch()
-
-        image = oci.Image("a:b", dest_dir)
-
-        # Create a dir tree to act as extracted "base"
-        # It contains a "first" dir and a "second" symlink pointing to "first"
-        rootfs_dir = tmp_path / "rootfs"
-        rootfs_dir.mkdir()
-
-        (rootfs_dir / "first").mkdir()
-        (rootfs_dir / "second").symlink_to("first")
-
-        assert len(temp_tar_contents) == 0
-        image.add_layer("tag", layer_dir, base_layer_dir=rootfs_dir)
-
-        # The tarfile must *not* contain the "./second" dir entry, to preserve
-        # the base layer symlink. Additionally, the file "second.txt" must
-        # be listed as inside "first/", and not "second/".
-        expected_tar_contents = [
-            "first",
-            "first/first.txt",
-            "first/second.txt",
-        ]
-        assert temp_tar_contents == expected_tar_contents
-
-    def test_add_layer_with_base_layer_dir_opaque(self, tmp_path, temp_tar_contents):
-        """
-        Test creating a layer with a base layer dir for reference, but the new
-        layer hides one of the dirs in the base layer via an opaque whiteout file.
-        """
-        dest_dir = tmp_path / "dest"
-        dest_dir.mkdir()
-
-        layer_dir = tmp_path / "layer_dir"
-        layer_dir.mkdir()
-
-        (layer_dir / "first").mkdir()
-        (layer_dir / "first/first.txt").touch()
-
-        (layer_dir / "second").mkdir()
-        (layer_dir / "second/second.txt").touch()
-        # Add an opaque marker to signify that "second/" should hide the
-        # base layer's "second" symlink.
-        overlays.oci_opaque_dir(layer_dir / "second").touch()
-
-        image = oci.Image("a:b", dest_dir)
-
-        # Create a dir tree to act as extracted "base"
-        # It contains a "first" dir and a "second" symlink pointing to "first"
-        rootfs_dir = tmp_path / "rootfs"
-        rootfs_dir.mkdir()
-
-        (rootfs_dir / "first").mkdir()
-        (rootfs_dir / "second").symlink_to("first")
-
-        assert len(temp_tar_contents) == 0
-        image.add_layer("tag", layer_dir, base_layer_dir=rootfs_dir)
-
-        # The tarfile *must* contain the "second" dir entry, because of the
-        # opaque whiteout file.
-        expected_tar_contents = [
-            "first",
-            "first/first.txt",
-            "second",
-            "second/.wh..wh..opq",
-            "second/second.txt",
-        ]
-        assert sorted(temp_tar_contents) == expected_tar_contents
-
-    def test_add_layer_with_base_layer_subdirs(self, tmp_path, temp_tar_contents):
-        """Test base layer handling with subdirectories."""
-        dest_dir = tmp_path / "dest"
-        dest_dir.mkdir()
-
-        layer_dir = tmp_path / "layer_dir"
-        layer_dir.mkdir()
-
-        # The path "second/" contains multiple subdirectories.
-        (layer_dir / "second").mkdir()
-        (layer_dir / "second/second.txt").touch()
-
-        (layer_dir / "second/subdir").mkdir()
-        (layer_dir / "second/subdir/subdir_file.txt").touch()
-
-        (layer_dir / "second/subdir/subsubdir").mkdir()
-        (layer_dir / "second/subdir/subsubdir/subsubdir_file.txt").touch()
-
-        (layer_dir / "third").mkdir()
-        (layer_dir / "third/third.txt").touch()
-
-        image = oci.Image("a:b", dest_dir)
-
-        rootfs_dir = tmp_path / "rootfs"
-        rootfs_dir.mkdir()
-
-        # The base layer has a "first/" dir and a symlink to it called "second".
-        # Every subdirectory in "second/" in the "upper" layer must be added as
-        # a subdir of "first".
-        (rootfs_dir / "first").mkdir()
-        (rootfs_dir / "second").symlink_to("first")
-
-        assert len(temp_tar_contents) == 0
-        image.add_layer("tag", layer_dir, base_layer_dir=rootfs_dir)
-
-        expected_tar_contents = [
-            "first/second.txt",
-            "first/subdir",
-            "first/subdir/subdir_file.txt",
-            "first/subdir/subsubdir",
-            "first/subdir/subsubdir/subsubdir_file.txt",
-            "third",
-            "third/third.txt",
-        ]
-        assert temp_tar_contents == expected_tar_contents
-
-    @staticmethod
-    def _duplicate_dirs_setup(tmp_path) -> Tuple[Path, Path]:
-        """Create a filetree with an upper layer and a fake 'rootfs' structure.
-
-        layer_dir/
-          |- bin/dir1/a.txt
-          |- usr/bin/dir1/b.txt
-        rootfs/
-          |- usr/bin/
-          |- bin ----> usr/bin (symlink)
-
-        returns a tuple with (layer_dir, rootfs).
-        """
-
-        layer_dir = tmp_path / "layer_dir"
-        layer_dir.mkdir()
-
-        # The new layer has "bin/dir1/a.txt" and "usr/bin/dir1/b.txt".
-        (layer_dir / "bin/dir1").mkdir(parents=True)
-        (layer_dir / "bin/dir1/a.txt").touch()
-        (layer_dir / "usr/bin/dir1").mkdir(parents=True)
-        (layer_dir / "usr/bin/dir1/b.txt").touch()
-
-        rootfs_dir = tmp_path / "rootfs"
-        rootfs_dir.mkdir()
-        # In the base layer "bin" is a symlink to "usr/bin"
-        (rootfs_dir / "usr/bin").mkdir(parents=True)
-        (rootfs_dir / "bin").symlink_to("usr/bin")
-
-        return layer_dir, rootfs_dir
-
-    def test_add_layer_duplicate_dirs(self, tmp_path, temp_tar_contents):
-        """
-        Test creating a layer where, because of symlinks in the base, multiple
-        directories end up as the same target.
-        """
-        dest_dir = tmp_path / "dest"
-        dest_dir.mkdir()
-
-        layer_dir, rootfs_dir = self._duplicate_dirs_setup(tmp_path)
-
-        image = oci.Image("a:b", dest_dir)
-        image.add_layer("tag", layer_dir, base_layer_dir=rootfs_dir)
-
-        expected_tar_contents = [
-            "usr",
-            "usr/bin",
-            "usr/bin/dir1",
-            "usr/bin/dir1/a.txt",
-            "usr/bin/dir1/b.txt",
-        ]
-        assert temp_tar_contents == expected_tar_contents
-
-    def test_add_layer_duplicate_dirs_conflict(self, tmp_path, temp_tar_contents):
-        """
-        Test creating a layer where, because of symlinks in the base, multiple
-        directories end up as the same target but the directories have different
-        ownership/permissions.
-        """
-        dest_dir = tmp_path / "dest"
-        dest_dir.mkdir()
-
-        layer_dir, rootfs_dir = self._duplicate_dirs_setup(tmp_path)
-
-        # Change the default permissions of the directories that will end up as
-        # "/usr/bin/dir1", to ensure that they are different.
-        (layer_dir / "bin/dir1").chmod(0o40770)
-        (layer_dir / "usr/bin/dir1").chmod(0o40775)
-
-        image = oci.Image("a:b", dest_dir)
-
-        # Check that the error message show the conflicting paths.
-        expected_message = (
-            "Conflicting paths pointing to 'usr/bin/dir1': "
-            f"{str(layer_dir / 'bin/dir1')}, "
-            f"{str(layer_dir / 'usr/bin/dir1')}"
-        )
-        with pytest.raises(errors.LayerArchivingError, match=expected_message):
-            image.add_layer("tag", layer_dir, base_layer_dir=rootfs_dir)
-
-    def test_add_layer_duplicate_files(self, tmp_path, temp_tar_contents):
-        """
-        Test creating a layer where, because of symlinks in the base, multiple
-        files (not directories) end up at the same target. The files have different
-        contents so this must raise an error.
-        """
-        dest_dir = tmp_path / "dest"
-        dest_dir.mkdir()
-
-        layer_dir, rootfs_dir = self._duplicate_dirs_setup(tmp_path)
-
-        # Create files with the same name but different contents in both
-        # directories.
-        (layer_dir / "bin/dir1/same.txt").write_text("first")
-        (layer_dir / "usr/bin/dir1/same.txt").write_text("second")
-
-        image = oci.Image("a:b", dest_dir)
-
-        # Check that the error message show the conflicting paths.
-        expected_message = (
-            "Conflicting paths pointing to 'usr/bin/dir1/same.txt': "
-            f"{str(layer_dir / 'bin/dir1/same.txt')}, "
-            f"{str(layer_dir / 'usr/bin/dir1/same.txt')}"
-        )
-        with pytest.raises(errors.LayerArchivingError, match=expected_message):
-            image.add_layer("tag", layer_dir, base_layer_dir=rootfs_dir)
-
     def test_add_new_user(
         self,
         check,
@@ -792,34 +477,6 @@ class TestImage:
                 (fake_tmp_new_layer / "etc/shadow").read_text(),
                 expected_user_files["shadow"],
             )
-
-    def test_add_layer_duplicate_identical_files(self, tmp_path, temp_tar_contents):
-        """
-        Test creating a layer where, because of symlinks in the base, multiple
-        files (not directories) end up at the same target. The files are identical
-        so the layer must be created successfully.
-        """
-        dest_dir = tmp_path / "dest"
-        dest_dir.mkdir()
-
-        layer_dir, rootfs_dir = self._duplicate_dirs_setup(tmp_path)
-
-        # Create files with the same contents with the same name in both directories
-        (layer_dir / "bin/dir1/same.txt").write_text("foobar")
-        (layer_dir / "usr/bin/dir1/same.txt").write_text("foobar")
-
-        image = oci.Image("a:b", dest_dir)
-        image.add_layer("tag", layer_dir, base_layer_dir=rootfs_dir)
-
-        expected_tar_contents = [
-            "usr",
-            "usr/bin",
-            "usr/bin/dir1",
-            "usr/bin/dir1/a.txt",
-            "usr/bin/dir1/b.txt",
-            "usr/bin/dir1/same.txt",
-        ]
-        assert temp_tar_contents == expected_tar_contents
 
     def test_to_docker_daemon(self, mock_run):
         image = oci.Image("a:b", Path("/c"))
@@ -1178,17 +835,6 @@ class TestImage:
             call(new_test_manifest_bytes),
             call(json.dumps(new_test_index).encode("utf-8")),
         ]
-
-    def test_archive_layer(self, mocker, new_dir):
-        Path("layer_dir").mkdir()
-        Path("layer_dir/bar.txt").touch()
-
-        spy_add = mocker.spy(tarfile.TarFile, "add")
-
-        oci._archive_layer(  # pylint: disable=protected-access
-            Path("layer_dir"), Path("./bar.tar")
-        )
-        assert spy_add.call_count == 1
 
     def test_stat(self, new_dir, mock_run, mocker):
         image_dir = Path("images/dir")
